Add files via upload
Ellen G. White's "Education" advocates for a holistic approach to learning, emphasizing the development of intellectual, moral, and spiritual dimensions. The book underscores the importance of integrating practical skills with academic knowledge, fostering character development, and viewing nature as a teacher. White's vision of education aims to prepare individuals for a life of service, aligning academic pursuits with ethical and spiritual growth.